requirements.txt
google-cloud-aiplatform==1.101.0
fastapi==0.115.14
qdrant-client==1.14.3
tqdm==4.67.1
uvicorn==0.35.0